feat: 초기 보스 & 고양이 공격 기능 구현
1. 보스가 고양이 공격하기 기능 구현 완료
2. 고양이가 보스 공격하기 기능 구현 완료

+++ Cat_Merge/Assets/1.Scripts/Contents/Mouse.cs
@@ -13,8 +13,8 @@ public class Mouse : MonoBehaviour
     private int mouseGrade;             // Áã µî±Þ (½ºÅ×ÀÌÁö)
     public int MouseGrade { get => mouseGrade; set => mouseGrade = value; }
 
-    private float mouseDamage;          // Áã ÀüÅõ·Â
-    public float MouseDamage { get => mouseDamage; set => mouseDamage = value; }
+    private int mouseDamage;            // Áã ÀüÅõ·Â
+    public int MouseDamage { get => mouseDamage; set => mouseDamage = value; }
 
     private float mouseHp;              // Áã Ã¼·Â
     public float MouseHp { get => mouseHp; set => mouseHp = value; }
@@ -25,7 +25,7 @@ public class Mouse : MonoBehaviour
     private int numOfAttack;            // °ø°Ý Å¸°Ù ¼ö
     public int NumOfAttack { get => numOfAttack; set => numOfAttack = value; }
 
-    public Mouse(int mouseId, string mouseName, int mouseGrade, float mouseDamage, float mouseHp, Sprite mouseImage, int numOfAttack)
+    public Mouse(int mouseId, string mouseName, int mouseGrade, int mouseDamage, float mouseHp, Sprite mouseImage, int numOfAttack)
     {
         MouseId = mouseId;
         MouseName = mouseName;
